some daniel changes
