+++ src/entries/popup/components/IconAndCopyList.tsx/IconAndCopyList.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
 
+import { goToNewTab } from '~/core/utils/tabs';
 import { Box, Inline, Separator, Stack, Symbol, Text } from '~/design-system';
 import { SymbolProps } from '~/design-system/components/Symbol/Symbol';
+import { TextLink } from '~/design-system/components/TextLink/TextLink';
 
 export interface IconAndCopyItem {
   icon: {
@@ -9,6 +11,10 @@ export interface IconAndCopyItem {
     color: SymbolProps['color'];
   };
   copy: string;
+  link?: {
+    hasLink: boolean;
+    href: string;
+  };
 }
 
 export function IconAndCopyList({
@@ -26,7 +32,7 @@ export function IconAndCopyList({
         </Box>
       }
     >
-      {iconAndCopyList?.map(({ icon, copy }) => (
+      {iconAndCopyList?.map(({ icon, copy, link }) => (
         <Box key={icon?.symbol}>
           <Inline wrap={false} space="16px" alignVertical="center">
             <Box display="flex" alignItems="center">
@@ -39,6 +45,16 @@ export function IconAndCopyList({
             </Box>
             <Text size="14pt" weight="medium" color="label">
               {copy}
+              {link && (
+                <TextLink
+                  color="blue"
+                  onClick={() =>
+                    goToNewTab({ url: 'https://learn.rainbow.me/' })
+                  }
+                >
+                  Learn More.
+                </TextLink>
+              )}
             </Text>
           </Inline>
         </Box>
